moved my page folder to the pages folder
