+++ scrybe_frontend/src/App.jsx
@@ -1,32 +1,32 @@
+import { Route, Routes } from "react-router-dom";
 import "./App.scss";
-import { Routes, Route } from "react-router-dom";
-import SentimentAnalysis from "./pages/SentimentAnalysis";
-import DummyPage from "./pages/DummyPage";
-import TranscribePage from "./pages/TranscribePage/TranscribePage";
-import { FAQs, TermsOfService } from "./pages";
-import LandingPage from "./pages/landing-page-a";
-import PageB from "./pages/Landing-page-b";
-import TermsAndCondition from "./pages/TermsAndCondition/TermsOfService";
-import History from "./pages/History/History";
-import Services from "./pages/services-page/Services";
-import Solutions from "./pages/solutions-page/Solutions";
-import DashboardOverview from "./pages/DashboardOverview";
-import Events from "./pages/Events-page/Events";
-import HelpSupport from "./pages/HelpAndSupport/HelpSupport";
-import General from "./components/HelpAndSupport/HelpAndSupportGeneral/General";
 import General1 from "./components/HelpAndSupport/general1/General1";
 import General2 from "./components/HelpAndSupport/general2/General2";
+import General from "./components/HelpAndSupport/HelpAndSupportGeneral/General";
 import PromotedArticles from "./components/HelpAndSupport/promoted-articles/PromotedArticles";
 import PromotedArticle1 from "./components/HelpAndSupport/PromotedArticle1/PromotedArticle1";
+import { FAQs, TermsOfService } from "./pages";
 import About from "./pages/AboutUs/About";
-import TryForFree from "./pages/TryForFree";
-import Industry from "./pages/indsutry/industry";
-import HowitWorks from "./pages/HowItWorks/HowItWorks";
-import UploadedRecordings from "./components/UploadRecordings";
 import Account from "./pages/Account";
-import Signin from "./pages/SignIn/SignIn";
 import Careers from "./pages/Careers";
+import DashboardOverview from "./pages/DashboardOverview";
+import DummyPage from "./pages/DummyPage";
+import Events from "./pages/Events-page/Events";
+import HelpSupport from "./pages/HelpAndSupport/HelpSupport";
+import History from "./pages/History/History";
+import HowitWorks from "./pages/HowItWorks/HowItWorks";
+import Industry from "./pages/indsutry/industry";
+import LandingPage from "./pages/landing-page-a";
+import PageB from "./pages/Landing-page-b";
 import PrivacyPolicy from "./pages/PrivacyPolicy";
+import SentimentAnalysis from "./pages/SentimentAnalysis";
+import Services from "./pages/services-page/Services";
+import Signin from "./pages/SignIn/SignIn";
+import Solutions from "./pages/solutions-page/Solutions";
+import TermsAndCondition from "./pages/TermsAndCondition/TermsOfService";
+import TranscribePage from "./pages/TranscribePage/TranscribePage";
+import TryForFree from "./pages/TryForFree";
+import UploadedRecordings from "./pages/UploadRecordings";
 
 function App() {
   return (
